% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSPlus.R
\name{ChangeEntrezFromName}
\alias{ChangeEntrezFromName}
\title{Changing the expression profile from Gene symbol to Entrez ID}
\usage{
ChangeEntrezFromName(GeneCount)
}
\arguments{
\item{GeneCount}{a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to gene symbols.}
}
\value{
a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to Entrez ID
}
\description{
Changing the expression profile from Gene symbol to Entrez ID
}
