% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSPlus.R
\name{DeleteDupChoseMax}
\alias{DeleteDupChoseMax}
\title{Deleting duplication genes}
\usage{
DeleteDupChoseMax(ExpDataframe, ReplaceName, by = "row", ori = FALSE)
}
\arguments{
\item{ExpDataframe}{a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to gene symbols.}

\item{ReplaceName}{a character with entrez id to be replaced.}

\item{by}{delete duplications by column or row.}

\item{ori}{keep original gene symbol.}
}
\value{
a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to entrez ID.
}
\description{
Deleting duplication genes
}
