% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSPlus.R
\name{PlotHeatMap}
\alias{PlotHeatMap}
\title{Title Ploting heatmap}
\usage{
PlotHeatMap(ClinData, GSVAData, CMSTME, CMS, SDI)
}
\arguments{
\item{ClinData}{a data frame with CMSPlus labels and CMS labels}

\item{GSVAData}{GSVA matrix}

\item{CMSTME}{CMSPlus label name}

\item{CMS}{CMS label name}

\item{SDI}{SDI label name}
}
\value{
a figure
}
\description{
Title Ploting heatmap
}
