% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSPlus.R
\name{RowDFtoRawNum}
\alias{RowDFtoRawNum}
\title{row to row}
\usage{
RowDFtoRawNum(ExpDataframe)
}
\arguments{
\item{ExpDataframe}{a dataframe with Gene Expression Profiles data values.}
}
\value{
a dataframe with Gene Expression Profiles data values
}
\description{
row to row
}
