% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSPlus.R
\name{CMSPlus}
\alias{CMSPlus}
\title{CMSPlus main function}
\usage{
CMSPlus(
  exp2symbol,
  prob = 0.5,
  plot = TRUE,
  CMSlabel = "RF.nearestCMS",
  CMSPluslabel = "nearest",
  parallel.sz = 0,
  InterGroupRandomize = T,
  seed = 123
)
}
\arguments{
\item{exp2symbol}{exp2symbol a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to gene symbols.}

\item{prob}{A numeric value between 0 and 1 specifying the minimum posterior probability threshold. Required to assign a predicted subtype. Samples with maximum subtype probability below this threshold will be classified as "unassigned". Default is 0.5.}

\item{plot}{TRUE produces plots; FALSE suppresses plotting.}

\item{CMSlabel}{A character string specifying the label used in plots to denote CMS subtypes.}

\item{CMSPluslabel}{A character string specifying the label used in plots to denote CMSPlus subtypes.}

\item{parallel.sz}{Integer specifying the number of parallel workers used for GSVA computation. Default is 1.}

\item{InterGroupRandomize}{Logical value indicating whether to perform randomization in plots.}

\item{seed}{Integer specifying the random seed used to ensure reproducibility of randomization step.}
}
\value{
list including CMSPlusLabel, CMSLabel, gsva_matrix and HeatmapPlot
}
\description{
CMSPlus main function
}
