% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSPlus.R
\name{PlotHeatMap}
\alias{PlotHeatMap}
\title{Ploting heatmap}
\usage{
PlotHeatMap(
  ClinData,
  GSVAData,
  CMSPlus,
  CMS,
  SDI,
  InterGroupRandomize = T,
  seed = 123
)
}
\arguments{
\item{ClinData}{a data frame with CMSPlus labels and CMS labels.}

\item{GSVAData}{GSVA matrix}

\item{CMSPlus}{CMSPlus label name}

\item{CMS}{CMS label name}

\item{SDI}{SDI label name}

\item{InterGroupRandomize}{TRUE or FALSE}

\item{seed}{numeric}
}
\value{
list including CMSPlusLabel, CMSLabel, gsva_matrix and HeatmapPlot
}
\description{
Ploting heatmap
}
